Update evaluator.py and plotter.py
