+++ autogluon_benchmark/evaluation/elo/elo_utils.py
@@ -1,3 +1,4 @@
+from collections import defaultdict
 from typing import List
 
 import math
@@ -6,6 +7,9 @@ import numpy as np
 import pandas as pd
 from sklearn.linear_model import LogisticRegression
 from tqdm import tqdm
+
+MODEL_A = "framework_1"
+MODEL_B = "framework_2"
 
 
 def compute_mle_elo(
@@ -206,3 +210,171 @@ def compute_elo_rating_dataset_contributon(
         elo_impact_by_dataset_list.append(delta)
     elo_impact_by_dataset = pd.concat(elo_impact_by_dataset_list, axis=1)
     return elo_impact_by_dataset
+
+
+def get_rank_confidence(df: pd.DataFrame):
+    df = df.copy()
+    df = df.sort_values(by=["Arena Elo"], ascending=False)
+
+    elo_ratings = df["Arena Elo"].to_list()
+    uppers = df["upper"].to_list()
+    lowers = df["lower"].to_list()
+
+    ranks = []
+
+    cur_rank = 0
+    prev_lower = None
+    num_models = len(elo_ratings)
+    for i in range(num_models):
+        cur_elo = elo_ratings[i]
+        cur_upper = uppers[i]
+        cur_lower = lowers[i]
+        if prev_lower is None or cur_upper < prev_lower:
+            cur_rank = i + 1
+            prev_lower = cur_lower
+        ranks.append(cur_rank)
+
+    df["Rank"] = ranks
+
+    return df
+
+
+def predict_win_rate(elo_ratings: dict, SCALE=400, BASE=10):
+    names = sorted(list(elo_ratings.keys()))
+    wins = defaultdict(lambda: defaultdict(lambda: 0))
+    for a in names:
+        for b in names:
+            ea = 1 / (1 + BASE ** ((elo_ratings[b] - elo_ratings[a]) / SCALE))
+            wins[a][b] = ea
+            wins[b][a] = 1 - ea
+
+    data = {
+        a: [wins[a][b] if a != b else np.NAN for b in names]
+        for a in names
+    }
+
+    df = pd.DataFrame(data, index=names)
+    df.index.name = "model_a"
+    df.columns.name = "model_b"
+    return df.T
+
+
+def _fix_missing(df, missing_A, missing_B):
+    df = df.copy()
+    for b in missing_B:
+        df[b] = 0
+    df = df.T
+    for a in missing_A:
+        df[a] = 0
+    df = df.T
+    return df
+
+
+def compute_pairwise_win_fraction(battles, max_num_models=30) -> pd.DataFrame:
+    unique_A = list(battles[MODEL_A].unique())
+    unique_B = list(battles[MODEL_B].unique())
+    missing_A = [b for b in unique_B if b not in unique_A]
+    missing_B = [a for a in unique_A if a not in unique_B]
+    unique_all = unique_A + missing_A
+    # Times each model wins as Model A
+    a_win_ptbl = pd.pivot_table(
+        battles[battles['winner'] == "1"],
+        index=MODEL_A, columns=MODEL_B, aggfunc="size", fill_value=0)
+
+    # Table counting times each model wins as Model B
+    b_win_ptbl = pd.pivot_table(
+        battles[battles['winner'] == "2"],
+        index=MODEL_A, columns=MODEL_B, aggfunc="size", fill_value=0)
+
+    # Table counting times each model wins as Model B
+    tie_ptbl = pd.pivot_table(
+        battles[battles['winner'] == "tie"],
+        index=MODEL_A, columns=MODEL_B, aggfunc="size", fill_value=0)
+    tie_ptbl *= 0.5
+
+    # Table counting number of A-B pairs
+    num_battles_ptbl = pd.pivot_table(battles,
+        index=MODEL_A, columns=MODEL_B, aggfunc="size", fill_value=0)
+
+    a_win_ptbl = _fix_missing(df=a_win_ptbl, missing_A=missing_A, missing_B=missing_B)
+    b_win_ptbl = _fix_missing(df=b_win_ptbl, missing_A=missing_A, missing_B=missing_B)
+    tie_missing_A = [a for a in unique_all if a not in tie_ptbl.index]
+    tie_missing_B = [b for b in unique_all if b not in tie_ptbl.columns]
+    tie_ptbl = _fix_missing(df=tie_ptbl, missing_A=tie_missing_A, missing_B=tie_missing_B)
+    num_battles_ptbl = _fix_missing(df=num_battles_ptbl, missing_A=missing_A, missing_B=missing_B)
+
+    # Computing the proportion of wins for each model as A and as B
+    # against all other models
+    row_beats_col_freq = (
+        (a_win_ptbl + b_win_ptbl.T + tie_ptbl + tie_ptbl.T) /
+        (num_battles_ptbl + num_battles_ptbl.T)
+    )
+
+    # Arrange ordering according to proprition of wins
+    prop_wins = row_beats_col_freq.mean(axis=1).sort_values(ascending=False)
+    prop_wins = prop_wins[:max_num_models]
+    model_names = list(prop_wins.keys())
+    row_beats_col = row_beats_col_freq.loc[model_names, model_names]
+    return row_beats_col
+
+
+def get_arena_leaderboard(bootstrap_elo_lu: pd.DataFrame, results_df: pd.DataFrame):
+    bars = pd.DataFrame(dict(
+        lower=bootstrap_elo_lu.quantile(.025),
+        rating=bootstrap_elo_lu.quantile(.5),
+        upper=bootstrap_elo_lu.quantile(.975))).reset_index(names="model").sort_values("rating", ascending=False)
+    bars['error_y'] = bars['upper'] - bars["rating"]
+    bars['error_y_minus'] = bars['rating'] - bars["lower"]
+    bars['rating_rounded'] = np.round(bars['rating'], 2)
+    battles = convert_results_to_battles(results_df=results_df)
+    from collections import defaultdict
+    framework_battle_counts = defaultdict(int)
+    framework_win_counts = defaultdict(float)
+    for f, win_val in [(MODEL_A, "1"), (MODEL_B, "2")]:
+        counts = battles[f].value_counts().to_dict()
+        win_counts = battles[[f, "winner"]].value_counts().reset_index()
+        win_counts.loc[win_counts["winner"] == "tie", "count"] *= 0.5
+        win_counts = win_counts.loc[win_counts["winner"].isin([win_val, "tie"]), :]
+        win_counts = win_counts.drop(columns=["winner"]).groupby(f)["count"].sum().to_dict()
+        for framework in counts:
+            framework_battle_counts[framework] += counts[framework]
+        for framework in win_counts:
+            framework_win_counts[framework] += win_counts[framework]
+    framework_battle_counts = dict(framework_battle_counts)
+
+    def _get_95_ci(upper, lower):
+        return f"+{upper:.0f}/-{lower:.0f}"
+
+    leaderboard = bars.copy()
+    leaderboard["95% CI"] = [_get_95_ci(upper, lower) for upper, lower in zip(leaderboard["error_y"], leaderboard["error_y_minus"])]
+    leaderboard["Arena Elo"] = np.round(leaderboard['rating'], 0).astype(int)
+    leaderboard["Battles"] = leaderboard["model"].map(framework_battle_counts)
+    leaderboard["Wins"] = np.round(leaderboard["model"].map(framework_win_counts), decimals=0).astype(int)
+    leaderboard["Winrate"] = np.round(leaderboard["Wins"] / leaderboard["Battles"], decimals=2)
+    leaderboard["Rank (Simple)"] = leaderboard["Arena Elo"].rank(method="min", ascending=False).astype(int)
+    leaderboard["Model"] = leaderboard["model"]
+    leaderboard = get_rank_confidence(df=leaderboard)
+
+    results_mean_agg = results_df[["framework", "rank", "bestdiff", "loss_rescaled"]].groupby("framework").mean()
+    leaderboard["mean_rank"] = leaderboard["model"].map(results_mean_agg["rank"])
+    leaderboard["mean_bestdiff"] = leaderboard["model"].map(results_mean_agg["bestdiff"])
+    leaderboard["mean_loss_rescaled"] = leaderboard["model"].map(results_mean_agg["loss_rescaled"])
+
+    leaderboard["Rank Avg"] = np.round(leaderboard["mean_rank"], decimals=1)
+    leaderboard["Champ Delta %"] = np.round(leaderboard["mean_bestdiff"] * 100, decimals=1)
+    leaderboard["Rescaled Loss"] = np.round(leaderboard["mean_loss_rescaled"], decimals=2)
+
+    leaderboard_print = leaderboard[[
+        "Rank",
+        "Model",
+        "Arena Elo",
+        "95% CI",
+        # "Battles",
+        # "Wins",
+        "Winrate",
+        "Rescaled Loss",
+        # "Rank Avg",
+        "Champ Delta %",
+    ]]
+
+    return leaderboard, leaderboard_print
